refactor: remove text field related generated classes and retain common functionality with a base class
Remove text field related generated classes. Move methods to the classes that use them. Retain common functionality with a base class. Update javadocs.
