--- vaadin-text-field-flow-parent/vaadin-text-field-flow/src/main/java/com/vaadin/flow/component/textfield/GeneratedVaadinTextField.java
@@ -1,1148 +1,0 @@
-/*
- * Copyright 2000-2022 Vaadin Ltd.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not
- * use this file except in compliance with the License. You may obtain a copy of
- * the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations under
- * the License.
- */
-package com.vaadin.flow.component.textfield;
-
-import com.vaadin.flow.component.AbstractSinglePropertyField;
-import com.vaadin.flow.component.Component;
-import com.vaadin.flow.component.ComponentEvent;
-import com.vaadin.flow.component.ComponentEventListener;
-import com.vaadin.flow.component.DomEvent;
-import com.vaadin.flow.component.Focusable;
-import com.vaadin.flow.component.HasStyle;
-import com.vaadin.flow.component.NotSupported;
-import com.vaadin.flow.component.Tag;
-import com.vaadin.flow.component.dependency.JsModule;
-import com.vaadin.flow.component.dependency.NpmPackage;
-import com.vaadin.flow.dom.Element;
-import com.vaadin.flow.function.SerializableBiFunction;
-import com.vaadin.flow.function.SerializableFunction;
-import com.vaadin.flow.shared.Registration;
-
-/**
- * <p>
- * Description copied from corresponding location in WebComponent:
- * </p>
- * <p>
- * {@code <vaadin-text-field>} is a Web Component for text field control in
- * forms.
- * </p>
- * <p>
- * &lt;vaadin-text-field label=&quot;First Name&quot;&gt;
- * &lt;/vaadin-text-field&gt;
- * </p>
- * <h3>Prefixes and suffixes</h3>
- * <p>
- * These are child elements of a {@code <vaadin-text-field>} that are displayed
- * inline with the input, before or after. In order for an element to be
- * considered as a prefix, it must have the slot attribute set to {@code prefix}
- * (and similarly for {@code suffix}).
- * </p>
- * <p>
- * &lt;vaadin-text-field label=&quot;Email address&quot;&gt; &lt;div
- * slot=&quot;prefix&quot;&gt;Sent to:&lt;/div&gt; &lt;div
- * slot=&quot;suffix&quot;&gt;@vaadin.com&lt;/div&gt; &lt;/vaadin-text-area&gt;
- * </p>
- * <h3>Styling</h3>
- * <p>
- * The following custom properties are available for styling:
- * </p>
- * <table>
- * <thead>
- * <tr>
- * <th>Custom property</th>
- * <th>Description</th>
- * <th>Default</th>
- * </tr>
- * </thead> <tbody>
- * <tr>
- * <td>{@code --vaadin-text-field-default-width}</td>
- * <td>Set the default width of the input field</td>
- * <td>{@code 12em}</td>
- * </tr>
- * </tbody>
- * </table>
- * <p>
- * The following shadow DOM parts are available for styling:
- * </p>
- * <table>
- * <thead>
- * <tr>
- * <th>Part name</th>
- * <th>Description</th>
- * </tr>
- * </thead> <tbody>
- * <tr>
- * <td>{@code label}</td>
- * <td>The label element</td>
- * </tr>
- * <tr>
- * <td>{@code input-field}</td>
- * <td>The element that wraps prefix, value and suffix</td>
- * </tr>
- * <tr>
- * <td>{@code value}</td>
- * <td>The text value element inside the {@code input-field} element</td>
- * </tr>
- * <tr>
- * <td>{@code error-message}</td>
- * <td>The error message element</td>
- * </tr>
- * </tbody>
- * </table>
- * <p>
- * The following state attributes are available for styling:
- * </p>
- * <table>
- * <thead>
- * <tr>
- * <th>Attribute</th>
- * <th>Description</th>
- * <th>Part name</th>
- * </tr>
- * </thead> <tbody>
- * <tr>
- * <td>{@code disabled}</td>
- * <td>Set to a disabled text field</td>
- * <td>:host</td>
- * </tr>
- * <tr>
- * <td>{@code has-value}</td>
- * <td>Set when the element has a value</td>
- * <td>:host</td>
- * </tr>
- * <tr>
- * <td>{@code has-label}</td>
- * <td>Set when the element has a label</td>
- * <td>:host</td>
- * </tr>
- * <tr>
- * <td>{@code invalid}</td>
- * <td>Set when the element is invalid</td>
- * <td>:host</td>
- * </tr>
- * <tr>
- * <td>{@code input-prevented}</td>
- * <td>Temporarily set when invalid input is prevented</td>
- * <td>:host</td>
- * </tr>
- * <tr>
- * <td>{@code focused}</td>
- * <td>Set when the element is focused</td>
- * <td>:host</td>
- * </tr>
- * <tr>
- * <td>{@code focus-ring}</td>
- * <td>Set when the element is keyboard focused</td>
- * <td>:host</td>
- * </tr>
- * <tr>
- * <td>{@code readonly}</td>
- * <td>Set to a readonly text field</td>
- * <td>:host</td>
- * </tr>
- * </tbody>
- * </table>
- * <p>
- * See
- * <a href="https://github.com/vaadin/vaadin-themable-mixin/wiki">ThemableMixin
- * – how to apply styles for shadow parts</a>
- * </p>
- */
-@Tag("vaadin-text-field")
-@NpmPackage(value = "@vaadin/polymer-legacy-adapter", version = "24.0.0-alpha2")
-@JsModule("@vaadin/polymer-legacy-adapter/style-modules.js")
-@NpmPackage(value = "@vaadin/text-field", version = "24.0.0-alpha2")
-@JsModule("@vaadin/text-field/src/vaadin-text-field.js")
-public abstract class GeneratedVaadinTextField<R extends GeneratedVaadinTextField<R, T>, T>
-        extends AbstractSinglePropertyField<R, T>
-        implements HasStyle, Focusable<R> {
-
-    /**
-     * <p>
-     * Description copied from corresponding location in WebComponent:
-     * </p>
-     * <p>
-     * Specify that this control should have input focus when the page loads.
-     * <p>
-     * This property is not synchronized automatically from the client side, so
-     * the returned value may not be the same as in client side.
-     * </p>
-     *
-     * @return the {@code autofocus} property from the webcomponent
-     */
-    protected boolean isAutofocusBoolean() {
-        return getElement().getProperty("autofocus", false);
-    }
-
-    /**
-     * <p>
-     * Description copied from corresponding location in WebComponent:
-     * </p>
-     * <p>
-     * Specify that this control should have input focus when the page loads.
-     * </p>
-     *
-     * @param autofocus
-     *            the boolean value to set
-     */
-    protected void setAutofocus(boolean autofocus) {
-        getElement().setProperty("autofocus", autofocus);
-    }
-
-    /**
-     * <p>
-     * Description copied from corresponding location in WebComponent:
-     * </p>
-     * <p>
-     * If true, the user cannot interact with this element.
-     * <p>
-     * This property is not synchronized automatically from the client side, so
-     * the returned value may not be the same as in client side.
-     * </p>
-     *
-     * @return the {@code disabled} property from the webcomponent
-     */
-    protected boolean isDisabledBoolean() {
-        return getElement().getProperty("disabled", false);
-    }
-
-    /**
-     * <p>
-     * Description copied from corresponding location in WebComponent:
-     * </p>
-     * <p>
-     * If true, the user cannot interact with this element.
-     * </p>
-     *
-     * @param disabled
-     *            the boolean value to set
-     */
-    protected void setDisabled(boolean disabled) {
-        getElement().setProperty("disabled", disabled);
-    }
-
-    /**
-     * <p>
-     * Description copied from corresponding location in WebComponent:
-     * </p>
-     * <p>
-     * Whether the value of the control can be automatically completed by the
-     * browser. List of available options at:
-     * https://developer.mozilla.org/en/docs
-     * /Web/HTML/Element/input#attr-autocomplete
-     * <p>
-     * This property is not synchronized automatically from the client side, so
-     * the returned value may not be the same as in client side.
-     * </p>
-     *
-     * @return the {@code autocomplete} property from the webcomponent
-     */
-    protected String getAutocompleteString() {
-        return getElement().getProperty("autocomplete");
-    }
-
-    /**
-     * <p>
-     * Description copied from corresponding location in WebComponent:
-     * </p>
-     * <p>
-     * Whether the value of the control can be automatically completed by the
-     * browser. List of available options at:
-     * https://developer.mozilla.org/en/docs
-     * /Web/HTML/Element/input#attr-autocomplete
-     * </p>
-     *
-     * @param autocomplete
-     *            the String value to set
-     */
-    protected void setAutocomplete(String autocomplete) {
-        getElement().setProperty("autocomplete",
-                autocomplete == null ? "" : autocomplete);
-    }
-
-    /**
-     * <p>
-     * Description copied from corresponding location in WebComponent:
-     * </p>
-     * <p>
-     * This is a property supported by Safari that is used to control whether
-     * autocorrection should be enabled when the user is entering/editing the
-     * text. Possible values are: on: Enable autocorrection. off: Disable
-     * autocorrection.
-     * <p>
-     * This property is not synchronized automatically from the client side, so
-     * the returned value may not be the same as in client side.
-     * </p>
-     *
-     * @return the {@code autocorrect} property from the webcomponent
-     */
-    protected String getAutocorrectString() {
-        return getElement().getProperty("autocorrect");
-    }
-
-    /**
-     * <p>
-     * Description copied from corresponding location in WebComponent:
-     * </p>
-     * <p>
-     * This is a property supported by Safari that is used to control whether
-     * autocorrection should be enabled when the user is entering/editing the
-     * text. Possible values are: on: Enable autocorrection. off: Disable
-     * autocorrection.
-     * </p>
-     *
-     * @param autocorrect
-     *            the String value to set
-     */
-    protected void setAutocorrect(String autocorrect) {
-        getElement().setProperty("autocorrect",
-                autocorrect == null ? "" : autocorrect);
-    }
-
-    /**
-     * <p>
-     * Description copied from corresponding location in WebComponent:
-     * </p>
-     * <p>
-     * This is a property supported by Safari and Chrome that is used to control
-     * whether autocapitalization should be enabled when the user is
-     * entering/editing the text. Possible values are: characters: Characters
-     * capitalization. words: Words capitalization. sentences: Sentences
-     * capitalization. none: No capitalization.
-     * <p>
-     * This property is not synchronized automatically from the client side, so
-     * the returned value may not be the same as in client side.
-     * </p>
-     *
-     * @return the {@code autocapitalize} property from the webcomponent
-     */
-    protected String getAutocapitalizeString() {
-        return getElement().getProperty("autocapitalize");
-    }
-
-    /**
-     * <p>
-     * Description copied from corresponding location in WebComponent:
-     * </p>
-     * <p>
-     * This is a property supported by Safari and Chrome that is used to control
-     * whether autocapitalization should be enabled when the user is
-     * entering/editing the text. Possible values are: characters: Characters
-     * capitalization. words: Words capitalization. sentences: Sentences
-     * capitalization. none: No capitalization.
-     * </p>
-     *
-     * @param autocapitalize
-     *            the String value to set
-     */
-    protected void setAutocapitalize(String autocapitalize) {
-        getElement().setProperty("autocapitalize",
-                autocapitalize == null ? "" : autocapitalize);
-    }
-
-    /**
-     * <p>
-     * Description copied from corresponding location in WebComponent:
-     * </p>
-     * <p>
-     * Specify that the value should be automatically selected when the field
-     * gains focus.
-     * <p>
-     * This property is not synchronized automatically from the client side, so
-     * the returned value may not be the same as in client side.
-     * </p>
-     *
-     * @return the {@code autoselect} property from the webcomponent
-     */
-    protected boolean isAutoselectBoolean() {
-        return getElement().getProperty("autoselect", false);
-    }
-
-    /**
-     * <p>
-     * Description copied from corresponding location in WebComponent:
-     * </p>
-     * <p>
-     * Specify that the value should be automatically selected when the field
-     * gains focus.
-     * </p>
-     *
-     * @param autoselect
-     *            the boolean value to set
-     */
-    protected void setAutoselect(boolean autoselect) {
-        getElement().setProperty("autoselect", autoselect);
-    }
-
-    /**
-     * <p>
-     * Description copied from corresponding location in WebComponent:
-     * </p>
-     * <p>
-     * Error to show when the input value is invalid.
-     * <p>
-     * This property is not synchronized automatically from the client side, so
-     * the returned value may not be the same as in client side.
-     * </p>
-     *
-     * @return the {@code errorMessage} property from the webcomponent
-     */
-    protected String getErrorMessageString() {
-        return getElement().getProperty("errorMessage");
-    }
-
-    /**
-     * <p>
-     * Description copied from corresponding location in WebComponent:
-     * </p>
-     * <p>
-     * Error to show when the input value is invalid.
-     * </p>
-     *
-     * @param errorMessage
-     *            the String value to set
-     */
-    protected void setErrorMessage(String errorMessage) {
-        getElement().setProperty("errorMessage",
-                errorMessage == null ? "" : errorMessage);
-    }
-
-    /**
-     * <p>
-     * Description copied from corresponding location in WebComponent:
-     * </p>
-     * <p>
-     * String used for the label element.
-     * <p>
-     * This property is not synchronized automatically from the client side, so
-     * the returned value may not be the same as in client side.
-     * </p>
-     *
-     * @return the {@code label} property from the webcomponent
-     */
-    protected String getLabelString() {
-        return getElement().getProperty("label");
-    }
-
-    /**
-     * <p>
-     * Description copied from corresponding location in WebComponent:
-     * </p>
-     * <p>
-     * String used for the label element.
-     * </p>
-     *
-     * @param label
-     *            the String value to set
-     */
-    protected void setLabel(String label) {
-        getElement().setProperty("label", label == null ? "" : label);
-    }
-
-    /**
-     * <p>
-     * Description copied from corresponding location in WebComponent:
-     * </p>
-     * <p>
-     * Maximum number of characters (in Unicode code points) that the user can
-     * enter.
-     * <p>
-     * This property is not synchronized automatically from the client side, so
-     * the returned value may not be the same as in client side.
-     * </p>
-     *
-     * @return the {@code maxlength} property from the webcomponent
-     */
-    protected double getMaxlengthDouble() {
-        return getElement().getProperty("maxlength", 0.0);
-    }
-
-    /**
-     * <p>
-     * Description copied from corresponding location in WebComponent:
-     * </p>
-     * <p>
-     * Maximum number of characters (in Unicode code points) that the user can
-     * enter.
-     * </p>
-     *
-     * @param maxlength
-     *            the double value to set
-     */
-    protected void setMaxlength(double maxlength) {
-        getElement().setProperty("maxlength", maxlength);
-    }
-
-    /**
-     * <p>
-     * Description copied from corresponding location in WebComponent:
-     * </p>
-     * <p>
-     * Minimum number of characters (in Unicode code points) that the user can
-     * enter.
-     * <p>
-     * This property is not synchronized automatically from the client side, so
-     * the returned value may not be the same as in client side.
-     * </p>
-     *
-     * @return the {@code minlength} property from the webcomponent
-     */
-    protected double getMinlengthDouble() {
-        return getElement().getProperty("minlength", 0.0);
-    }
-
-    /**
-     * <p>
-     * Description copied from corresponding location in WebComponent:
-     * </p>
-     * <p>
-     * Minimum number of characters (in Unicode code points) that the user can
-     * enter.
-     * </p>
-     *
-     * @param minlength
-     *            the double value to set
-     */
-    protected void setMinlength(double minlength) {
-        getElement().setProperty("minlength", minlength);
-    }
-
-    /**
-     * <p>
-     * Description copied from corresponding location in WebComponent:
-     * </p>
-     * <p>
-     * The name of the control, which is submitted with the form data.
-     * <p>
-     * This property is not synchronized automatically from the client side, so
-     * the returned value may not be the same as in client side.
-     * </p>
-     *
-     * @return the {@code name} property from the webcomponent
-     */
-    protected String getNameString() {
-        return getElement().getProperty("name");
-    }
-
-    /**
-     * <p>
-     * Description copied from corresponding location in WebComponent:
-     * </p>
-     * <p>
-     * The name of the control, which is submitted with the form data.
-     * </p>
-     *
-     * @param name
-     *            the String value to set
-     */
-    protected void setName(String name) {
-        getElement().setProperty("name", name == null ? "" : name);
-    }
-
-    /**
-     * <p>
-     * Description copied from corresponding location in WebComponent:
-     * </p>
-     * <p>
-     * A hint to the user of what can be entered in the control.
-     * <p>
-     * This property is not synchronized automatically from the client side, so
-     * the returned value may not be the same as in client side.
-     * </p>
-     *
-     * @return the {@code placeholder} property from the webcomponent
-     */
-    protected String getPlaceholderString() {
-        return getElement().getProperty("placeholder");
-    }
-
-    /**
-     * <p>
-     * Description copied from corresponding location in WebComponent:
-     * </p>
-     * <p>
-     * A hint to the user of what can be entered in the control.
-     * </p>
-     *
-     * @param placeholder
-     *            the String value to set
-     */
-    protected void setPlaceholder(String placeholder) {
-        getElement().setProperty("placeholder",
-                placeholder == null ? "" : placeholder);
-    }
-
-    /**
-     * <p>
-     * Description copied from corresponding location in WebComponent:
-     * </p>
-     * <p>
-     * This attribute indicates that the user cannot modify the value of the
-     * control.
-     * <p>
-     * This property is not synchronized automatically from the client side, so
-     * the returned value may not be the same as in client side.
-     * </p>
-     *
-     * @return the {@code readonly} property from the webcomponent
-     */
-    protected boolean isReadonlyBoolean() {
-        return getElement().getProperty("readonly", false);
-    }
-
-    /**
-     * <p>
-     * Description copied from corresponding location in WebComponent:
-     * </p>
-     * <p>
-     * This attribute indicates that the user cannot modify the value of the
-     * control.
-     * </p>
-     *
-     * @param readonly
-     *            the boolean value to set
-     */
-    protected void setReadonly(boolean readonly) {
-        getElement().setProperty("readonly", readonly);
-    }
-
-    /**
-     * <p>
-     * Description copied from corresponding location in WebComponent:
-     * </p>
-     * <p>
-     * Specifies that the user must fill in a value.
-     * <p>
-     * This property is not synchronized automatically from the client side, so
-     * the returned value may not be the same as in client side.
-     * </p>
-     *
-     * @return the {@code required} property from the webcomponent
-     */
-    protected boolean isRequiredBoolean() {
-        return getElement().getProperty("required", false);
-    }
-
-    /**
-     * <p>
-     * Description copied from corresponding location in WebComponent:
-     * </p>
-     * <p>
-     * Specifies that the user must fill in a value.
-     * </p>
-     *
-     * @param required
-     *            the boolean value to set
-     */
-    protected void setRequired(boolean required) {
-        getElement().setProperty("required", required);
-    }
-
-    /**
-     * <p>
-     * Description copied from corresponding location in WebComponent:
-     * </p>
-     * <p>
-     * This property is set to true when the control value is invalid.
-     * </p>
-     *
-     * @return the {@code invalid} property from the webcomponent
-     */
-    protected boolean isInvalidBoolean() {
-        return getElement().getProperty("invalid", false);
-    }
-
-    /**
-     * <p>
-     * Description copied from corresponding location in WebComponent:
-     * </p>
-     * <p>
-     * This property is set to true when the control value is invalid.
-     * </p>
-     *
-     * @param invalid
-     *            the boolean value to set
-     */
-    protected void setInvalid(boolean invalid) {
-        getElement().setProperty("invalid", invalid);
-    }
-
-    /**
-     * <p>
-     * Description copied from corresponding location in WebComponent:
-     * </p>
-     * <p>
-     * Identifies a list of pre-defined options to suggest to the user. The
-     * value must be the id of a <datalist> element in the same document.
-     * <p>
-     * This property is not synchronized automatically from the client side, so
-     * the returned value may not be the same as in client side.
-     * </p>
-     *
-     * @return the {@code list} property from the webcomponent
-     */
-    protected String getListString() {
-        return getElement().getProperty("list");
-    }
-
-    /**
-     * <p>
-     * Description copied from corresponding location in WebComponent:
-     * </p>
-     * <p>
-     * Identifies a list of pre-defined options to suggest to the user. The
-     * value must be the id of a <datalist> element in the same document.
-     * </p>
-     *
-     * @param list
-     *            the String value to set
-     */
-    protected void setList(String list) {
-        getElement().setProperty("list", list == null ? "" : list);
-    }
-
-    /**
-     * <p>
-     * Description copied from corresponding location in WebComponent:
-     * </p>
-     * <p>
-     * A regular expression that the value is checked against. The pattern must
-     * match the entire value, not just some subset.
-     * <p>
-     * This property is not synchronized automatically from the client side, so
-     * the returned value may not be the same as in client side.
-     * </p>
-     *
-     * @return the {@code pattern} property from the webcomponent
-     */
-    protected String getPatternString() {
-        return getElement().getProperty("pattern");
-    }
-
-    /**
-     * <p>
-     * Description copied from corresponding location in WebComponent:
-     * </p>
-     * <p>
-     * A regular expression that the value is checked against. The pattern must
-     * match the entire value, not just some subset.
-     * </p>
-     *
-     * @param pattern
-     *            the String value to set
-     */
-    protected void setPattern(String pattern) {
-        getElement().setProperty("pattern", pattern == null ? "" : pattern);
-    }
-
-    /**
-     * <p>
-     * Description copied from corresponding location in WebComponent:
-     * </p>
-     * <p>
-     * Message to show to the user when validation fails.
-     * <p>
-     * This property is not synchronized automatically from the client side, so
-     * the returned value may not be the same as in client side.
-     * </p>
-     *
-     * @return the {@code title} property from the webcomponent
-     */
-    protected String getTitleString() {
-        return getElement().getProperty("title");
-    }
-
-    /**
-     * <p>
-     * Description copied from corresponding location in WebComponent:
-     * </p>
-     * <p>
-     * The text usually displayed in a tooltip popup when the mouse is over the
-     * field.
-     * </p>
-     *
-     * @param title
-     *            the String value to set
-     */
-    protected void setTitle(String title) {
-        getElement().setProperty("title", title == null ? "" : title);
-    }
-
-    /**
-     * <p>
-     * Description copied from corresponding location in WebComponent:
-     * </p>
-     * <p>
-     * Returns true if the current input value satisfies all constraints (if
-     * any)
-     * </p>
-     * <p>
-     * This function is not supported by Flow because it returns a
-     * <code>boolean</code>. Functions with return types different than void are
-     * not supported at this moment.
-     */
-    @NotSupported
-    protected void checkValidity() {
-    }
-
-    /**
-     * <p>
-     * Description copied from corresponding location in WebComponent:
-     * </p>
-     * <p>
-     * Returns true if {@code value} is valid. {@code <iron-form>} uses this to
-     * check the validity or all its elements.
-     * </p>
-     * <p>
-     * This function is not supported by Flow because it returns a
-     * <code>boolean</code>. Functions with return types different than void are
-     * not supported at this moment.
-     */
-    @NotSupported
-    protected void validate() {
-    }
-
-    @DomEvent("change")
-    public static class ChangeEvent<R extends GeneratedVaadinTextField<R, ?>>
-            extends ComponentEvent<R> {
-        public ChangeEvent(R source, boolean fromClient) {
-            super(source, fromClient);
-        }
-    }
-
-    /**
-     * Adds a listener for {@code change} events fired by the webcomponent.
-     *
-     * @param listener
-     *            the listener
-     * @return a {@link Registration} for removing the event listener
-     */
-    @SuppressWarnings({ "rawtypes", "unchecked" })
-    protected Registration addChangeListener(
-            ComponentEventListener<ChangeEvent<R>> listener) {
-        return addListener(ChangeEvent.class,
-                (ComponentEventListener) listener);
-    }
-
-    public static class InvalidChangeEvent<R extends GeneratedVaadinTextField<R, ?>>
-            extends ComponentEvent<R> {
-        private final boolean invalid;
-
-        public InvalidChangeEvent(R source, boolean fromClient) {
-            super(source, fromClient);
-            this.invalid = source.isInvalidBoolean();
-        }
-
-        public boolean isInvalid() {
-            return invalid;
-        }
-    }
-
-    /**
-     * Adds a listener for {@code invalid-changed} events fired by the
-     * webcomponent.
-     *
-     * @param listener
-     *            the listener
-     * @return a {@link Registration} for removing the event listener
-     */
-    protected Registration addInvalidChangeListener(
-            ComponentEventListener<InvalidChangeEvent<R>> listener) {
-        return getElement()
-                .addPropertyChangeListener("invalid",
-                        event -> listener.onComponentEvent(
-                                new InvalidChangeEvent<R>((R) this,
-                                        event.isUserOriginated())));
-    }
-
-    /**
-     * Adds the given components as children of this component at the slot
-     * 'prefix'.
-     *
-     * @param components
-     *            The components to add.
-     * @see <a href=
-     *      "https://developer.mozilla.org/en-US/docs/Web/HTML/Element/slot">MDN
-     *      page about slots</a>
-     * @see <a href=
-     *      "https://html.spec.whatwg.org/multipage/scripting.html#the-slot-element">Spec
-     *      website about slots</a>
-     */
-    protected void addToPrefix(Component... components) {
-        for (Component component : components) {
-            component.getElement().setAttribute("slot", "prefix");
-            getElement().appendChild(component.getElement());
-        }
-    }
-
-    /**
-     * Adds the given components as children of this component at the slot
-     * 'input'.
-     *
-     * @param components
-     *            The components to add.
-     * @see <a href=
-     *      "https://developer.mozilla.org/en-US/docs/Web/HTML/Element/slot">MDN
-     *      page about slots</a>
-     * @see <a href=
-     *      "https://html.spec.whatwg.org/multipage/scripting.html#the-slot-element">Spec
-     *      website about slots</a>
-     */
-    protected void addToInput(Component... components) {
-        for (Component component : components) {
-            component.getElement().setAttribute("slot", "input");
-            getElement().appendChild(component.getElement());
-        }
-    }
-
-    /**
-     * Adds the given components as children of this component at the slot
-     * 'suffix'.
-     *
-     * @param components
-     *            The components to add.
-     * @see <a href=
-     *      "https://developer.mozilla.org/en-US/docs/Web/HTML/Element/slot">MDN
-     *      page about slots</a>
-     * @see <a href=
-     *      "https://html.spec.whatwg.org/multipage/scripting.html#the-slot-element">Spec
-     *      website about slots</a>
-     */
-    protected void addToSuffix(Component... components) {
-        for (Component component : components) {
-            component.getElement().setAttribute("slot", "suffix");
-            getElement().appendChild(component.getElement());
-        }
-    }
-
-    /**
-     * Removes the given child components from this component.
-     *
-     * @param components
-     *            The components to remove.
-     * @throws IllegalArgumentException
-     *             if any of the components is not a child of this component.
-     */
-    protected void remove(Component... components) {
-        for (Component component : components) {
-            if (getElement().equals(component.getElement().getParent())) {
-                component.getElement().removeAttribute("slot");
-                getElement().removeChild(component.getElement());
-            } else {
-                throw new IllegalArgumentException("The given component ("
-                        + component + ") is not a child of this component");
-            }
-        }
-    }
-
-    /**
-     * Removes all contents from this component, this includes child components,
-     * text content as well as child elements that have been added directly to
-     * this component using the {@link Element} API.
-     */
-    protected void removeAll() {
-        getElement().getChildren()
-                .forEach(child -> child.removeAttribute("slot"));
-        getElement().removeAllChildren();
-    }
-
-    /**
-     * Constructs a new GeneratedVaadinTextField component with the given
-     * arguments.
-     *
-     * @param initialValue
-     *            the initial value to set to the value
-     * @param defaultValue
-     *            the default value to use if the value isn't defined
-     * @param elementPropertyType
-     *            the type of the element property
-     * @param presentationToModel
-     *            a function that converts a string value to a model value
-     * @param modelToPresentation
-     *            a function that converts a model value to a string value
-     * @param <P>
-     *            the property type
-     */
-    public <P> GeneratedVaadinTextField(T initialValue, T defaultValue,
-            Class<P> elementPropertyType,
-            SerializableFunction<P, T> presentationToModel,
-            SerializableFunction<T, P> modelToPresentation) {
-        this(initialValue, defaultValue, elementPropertyType,
-                presentationToModel, modelToPresentation, false);
-    }
-
-    /**
-     * Constructs a new GeneratedVaadinTextField component with the given
-     * arguments.
-     * <p>
-     * If {@code isInitialValueOptional} is {@code true} then the initial value
-     * is used only if element has no {@code "value"} property value, otherwise
-     * element {@code "value"} property is ignored and the initial value is set.
-     *
-     * @param initialValue
-     *            the initial value to set to the value
-     * @param defaultValue
-     *            the default value to use if the value isn't defined
-     * @param elementPropertyType
-     *            the type of the element property
-     * @param presentationToModel
-     *            a function that converts a string value to a model value
-     * @param modelToPresentation
-     *            a function that converts a model value to a string value
-     *
-     * @param isInitialValueOptional
-     *            if {@code isInitialValueOptional} is {@code true} then the
-     *            initial value is used only if element has no {@code "value"}
-     *            property value, otherwise element {@code "value"} property is
-     *            ignored and the initial value is set
-     * @param <P>
-     *            the property type
-     */
-    public <P> GeneratedVaadinTextField(T initialValue, T defaultValue,
-            Class<P> elementPropertyType,
-            SerializableFunction<P, T> presentationToModel,
-            SerializableFunction<T, P> modelToPresentation,
-            boolean isInitialValueOptional) {
-        super("value", defaultValue, elementPropertyType, presentationToModel,
-                modelToPresentation);
-        if ((getElement().getProperty("value") == null
-                || !isInitialValueOptional) && initialValue != null) {
-            setPresentationValue(initialValue);
-        }
-    }
-
-    /**
-     * Constructs a new GeneratedVaadinTextField component with the given
-     * arguments.
-     *
-     * @param initialValue
-     *            the initial value to set to the value
-     * @param defaultValue
-     *            the default value to use if the value isn't defined
-     * @param acceptNullValues
-     *            whether <code>null</code> is accepted as a model value
-     */
-    public GeneratedVaadinTextField(T initialValue, T defaultValue,
-            boolean acceptNullValues) {
-        this(initialValue, defaultValue, acceptNullValues, false);
-    }
-
-    /**
-     * Constructs a new GeneratedVaadinTextField component with the given
-     * arguments.
-     * <p>
-     * If {@code isInitialValueOptional} is {@code true} then the initial value
-     * is used only if element has no {@code "value"} property value, otherwise
-     * element {@code "value"} property is ignored and the initial value is set.
-     *
-     * @param initialValue
-     *            the initial value to set to the value
-     * @param defaultValue
-     *            the default value to use if the value isn't defined
-     * @param acceptNullValues
-     *            whether <code>null</code> is accepted as a model value
-     * @param isInitialValueOptional
-     *            if {@code isInitialValueOptional} is {@code true} then the
-     *            initial value is used only if element has no {@code "value"}
-     *            property value, otherwise element {@code "value"} property is
-     *            ignored and the initial value is set
-     */
-    public GeneratedVaadinTextField(T initialValue, T defaultValue,
-            boolean acceptNullValues, boolean isInitialValueOptional) {
-        super("value", defaultValue, acceptNullValues);
-        if ((getElement().getProperty("value") == null
-                || !isInitialValueOptional) && initialValue != null) {
-            setPresentationValue(initialValue);
-        }
-    }
-
-    /**
-     * Constructs a new GeneratedVaadinTextField component with the given
-     * arguments.
-     *
-     * @param initialValue
-     *            the initial value to set to the value
-     * @param defaultValue
-     *            the default value to use if the value isn't defined
-     * @param elementPropertyType
-     *            the type of the element property
-     * @param presentationToModel
-     *            a function that accepts this component and a property value
-     *            and returns a model value
-     * @param modelToPresentation
-     *            a function that accepts this component and a model value and
-     *            returns a property value
-     * @param <P>
-     *            the property type
-     */
-    public <P> GeneratedVaadinTextField(T initialValue, T defaultValue,
-            Class<P> elementPropertyType,
-            SerializableBiFunction<R, P, T> presentationToModel,
-            SerializableBiFunction<R, T, P> modelToPresentation) {
-        this(initialValue, defaultValue, elementPropertyType,
-                presentationToModel, modelToPresentation, false);
-    }
-
-    /**
-     * Constructs a new GeneratedVaadinTextField component with the given
-     * arguments.
-     * <p>
-     * If {@code isInitialValueOptional} is {@code true} then the initial value
-     * is used only if element has no {@code "value"} property value, otherwise
-     * element {@code "value"} property is ignored and the initial value is set.
-     *
-     *
-     * @param initialValue
-     *            the initial value to set to the value
-     * @param defaultValue
-     *            the default value to use if the value isn't defined
-     * @param elementPropertyType
-     *            the type of the element property
-     * @param presentationToModel
-     *            a function that accepts this component and a property value
-     *            and returns a model value
-     * @param modelToPresentation
-     *            a function that accepts this component and a model value and
-     *            returns a property value
-     * @param isInitialValueOptional
-     *            if {@code isInitialValueOptional} is {@code true} then the
-     *            initial value is used only if element has no {@code "value"}
-     *            property value, otherwise element {@code "value"} property is
-     *            ignored and the initial value is set
-     * @param <P>
-     *            the property type
-     */
-    public <P> GeneratedVaadinTextField(T initialValue, T defaultValue,
-            Class<P> elementPropertyType,
-            SerializableBiFunction<R, P, T> presentationToModel,
-            SerializableBiFunction<R, T, P> modelToPresentation,
-            boolean isInitialValueOptional) {
-        super("value", defaultValue, elementPropertyType, presentationToModel,
-                modelToPresentation);
-        if ((getElement().getProperty("value") == null
-                || !isInitialValueOptional) && initialValue != null) {
-            setPresentationValue(initialValue);
-        }
-    }
-
-    /**
-     * Default constructor.
-     */
-    public GeneratedVaadinTextField() {
-        this(null, null, null, (SerializableFunction) null,
-                (SerializableFunction) null);
-    }
-}
